requirements.txt
streamlit==1.33.0
openai==1.17.1
pypdf==4.2.0
tiktoken==0.6.0
faiss-cpu==1.8.0
python-dotenv==1.0.1
httpx==0.27.0 